+++ server/service.go
@@ -1,0 +1,11 @@
+package server
+
+import (
+	pb "bms/bmsproto"
+	"bms/database"
+)
+
+type BmsServer struct {
+	pb.UnimplementedBmsDatabaseCrudServer
+	Db database.DataBase
+}
